refactor: abstract out useSingleHopChannelInfo
